Test: S3 파일 삭제 기능 테스트

+++ src/main/java/com/auction/usedauction/util/S3FileUploader.java
@@ -7,6 +7,7 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.auction.usedauction.UploadFIleDTO;
 import com.auction.usedauction.exception.FileEmptyException;
+import com.auction.usedauction.exception.S3FileNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -65,6 +66,10 @@ public class S3FileUploader {
     }
 
     public String deleteFile(String filePath) {
+        boolean isExist = amazonS3Client.doesObjectExist(bucket, filePath);
+        if (!isExist) {
+            throw new S3FileNotFoundException("S3에서 해당 파일을 찾지 못했습니다.");
+        }
         amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, filePath));
         log.info("S3 파일 삭제 완료 deletedFilePath = {}", filePath);
         return filePath;
